+++ app/Http/Controllers/Mazad/RoomController.php
@@ -447,7 +447,7 @@ class RoomController extends Controller
     {
         $room = MazadRoom::where('code', $code)->firstOrFail();
 
-        if ($room->owner_id !== Auth::id()) {
+        if ((int)$room->owner_id !== (int)Auth::id()) {
             return response()->json(['message' => 'Only the room owner can close the room'], 403);
         }
 
@@ -471,7 +471,7 @@ class RoomController extends Controller
     {
         $room = MazadRoom::where('code', $code)->firstOrFail();
 
-        if ($room->owner_id !== Auth::id()) {
+        if ((int)$room->owner_id !== (int)Auth::id()) {
             return response()->json(['message' => 'Only the room owner can assign teams'], 403);
         }
 
@@ -517,7 +517,7 @@ class RoomController extends Controller
     {
         $room = MazadRoom::where('code', $code)->firstOrFail();
 
-        if ($room->owner_id !== Auth::id()) {
+        if ((int)$room->owner_id !== (int)Auth::id()) {
             return response()->json(['message' => 'Only the room owner can assign teams'], 403);
         }
